seperated functional component form display components and added functionality for addToBasket button
